pyproject.toml
[tool.poetry]
name = "arai"
version = "3.1.0"
description = ""
authors = ["NimajnebEC <nimajnebec@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
"discord.py" = "^2.0.1"
aiomysql = "^0.1.1"
colouredlogs = "^10.0.1"


[tool.poetry.group.dev.dependencies]
black = "^23.9.1"
flake8 = "^6.1.0"
flake8-length = "^0.3.1"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
